Create Ultimate Badass Mech
Unfinished, so tl;dr I used AI to help fix an old Holo Mech E2, my E2 was using incorrect syntax for the newest Expression 2 in Wiremod, it started writing from this simple code I had, into a 150 part code that turned into this monstrosity. This is only 130 parts of it before the AI froze, I'm not sure if I can recreate it but will attempt regardless.